fix abstract/liberate calls
