+++ Cargo.toml
@@ -22,9 +22,9 @@ itertools = "0.10"
 psfparser = { git = "https://github.com/rahulk29/psfparser.git", branch = "master" }
 arcstr = "1"
 anyhow = "1"
-calibre = { git = "https://github.com/rahulk29/calibre", branch = "main", optional = true }
-abstract_lef = { git = "https://github.com/rahulk29/abstract-lef.git", branch = "main", optional = true }
-liberate_mx = { git = "https://github.com/rahulk29/liberate-mx.git", branch = "master", optional = true }
+calibre = { git = "ssh://git@bwrcrepo.eecs.berkeley.edu/rahulkumar/calibre.git", branch = "master", optional = true }
+abstract_lef = { git = "ssh://git@bwrcrepo.eecs.berkeley.edu/rahulkumar/abstract_lef.git", branch = "master", optional = true }
+liberate_mx = { git = "ssh://git@bwrcrepo.eecs.berkeley.edu/rahulkumar/liberate-mx.git", branch = "master", optional = true }
 tera = "1"
 lazy_static = "1"
 colored = "2"
@@ -38,17 +38,14 @@ sky130_open_pdk = { git = "https://github.com/substrate-labs/substrate", branch 
 ngspice = { git = "https://github.com/substrate-labs/substrate", branch = "main" }
 sky130_commercial_pdk = { git = "https://github.com/substrate-labs/substrate", branch = "main", optional = true }
 spectre = { git = "https://github.com/substrate-labs/substrate", branch = "main", optional = true }
-sub_calibre = { git = "https://github.com/rahulk29/substrate_mentor_plugins", branch = "main", optional = true }
+sub_calibre = { git = "ssh://git@bwrcrepo.eecs.berkeley.edu/rohankumar/substrate_mentor_plugins.git", branch = "main", optional = true }
 
 [dev-dependencies]
 tempfile = "3"
 paste = "1"
 
 [features]
-commercial = ["dep:calibre", "calibre", "dep:abstract_lef", "abstract_lef", "dep:liberate_mx", "dep:spectre", "dep:sky130_commercial_pdk", "dep:sub_calibre", "sub_calibre",]
-abstract_lef = ["abstract_lef/commercial"]
-sub_calibre = ["sub_calibre/commercial"]
-calibre = ["calibre/commercial"]
+commercial = ["dep:calibre", "dep:abstract_lef", "dep:liberate_mx", "dep:spectre", "dep:sky130_commercial_pdk", "dep:sub_calibre"]
 
 default = ["commercial"]
